+++ bricks/diagram/src/draw-canvas/constants.ts
@@ -1,0 +1,7 @@
+export const SYMBOL_FOR_SIZE_INITIALIZED = Symbol.for("size-initialized");
+export const DEFAULT_NODE_SIZE = 20;
+export const DEFAULT_NODE_GAP = 36;
+export const DEFAULT_AREA_WIDTH = 100;
+export const DEFAULT_AREA_HEIGHT = 60;
+export const DEFAULT_SCALE_RANGE_MIN = 0.5;
+export const DEFAULT_SCALE_RANGE_MAX = 2;