+++ bricks/icons/src/general-icon/index.spec.ts
@@ -15,11 +15,20 @@ jest.mock("../easyops-icon/index.js", () => ({
   },
 }));
 
-jest.mock("../fa-icon/index.js", () => ({
-  WrappedFaIcon(props: unknown) {
-    return ` fa-icon:${JSON.stringify(props)}`;
-  },
-}));
+jest.mock("../fa-icon/index.js", () => {
+  const React = jest.requireActual("react") as any;
+  return {
+    WrappedFaIcon(props: any) {
+      const { icon, onIconFoundChange } = props;
+      React.useEffect(() => {
+        if (icon === "unknown") {
+          onIconFoundChange?.(new CustomEvent("icon.found", { detail: false }));
+        }
+      }, [icon, onIconFoundChange]);
+      return ` fa-icon:${JSON.stringify(props)}`;
+    },
+  };
+});
 
 jest.mock("../img-icon/index.js", () => ({
   WrappedEoImgIcon(props: unknown) {
@@ -86,5 +95,84 @@ describe("eo-icon", () => {
       document.body.removeChild(element);
     });
     expect(element.shadowRoot?.textContent).toMatchInlineSnapshot(`""`);
+  });
+
+  test("unknown icon with fallback", async () => {
+    const element = document.createElement("eo-icon") as GeneralIcon;
+    element.lib = "fa";
+    element.icon = "unknown";
+    element.fallback = {
+      lib: "antd",
+      icon: "x",
+    };
+
+    act(() => {
+      document.body.appendChild(element);
+    });
+    expect(element.shadowRoot?.textContent).toMatchInlineSnapshot(
+      `"styles.shadow.css antd-icon:{"icon":"x"}"`
+    );
+
+    act(() => {
+      document.body.removeChild(element);
+    });
+  });
+
+  test("unknown lib with fallback", async () => {
+    const element = document.createElement("eo-icon") as GeneralIcon;
+    element.lib = "oops" as any;
+    element.icon = "any";
+    element.fallback = {
+      lib: "antd",
+      icon: "x",
+    };
+
+    act(() => {
+      document.body.appendChild(element);
+    });
+    expect(element.shadowRoot?.textContent).toMatchInlineSnapshot(
+      `"styles.shadow.css antd-icon:{"icon":"x"}"`
+    );
+
+    act(() => {
+      document.body.removeChild(element);
+    });
+  });
+
+  test("unknown lib with svg fallback", async () => {
+    const element = document.createElement("eo-icon") as GeneralIcon;
+    element.lib = "oops" as any;
+    element.icon = "any";
+    element.fallback = {
+      imgSrc: "https://test.com/image.svg",
+    };
+
+    act(() => {
+      document.body.appendChild(element);
+    });
+    expect(element.shadowRoot?.textContent).toMatchInlineSnapshot(
+      `"styles.shadow.css svg-icon:{"imgSrc":"https://test.com/image.svg"}"`
+    );
+
+    act(() => {
+      document.body.removeChild(element);
+    });
+  });
+
+  test("unknown lib without fallback", async () => {
+    const element = document.createElement("eo-icon") as GeneralIcon;
+    element.lib = "oops" as any;
+    element.icon = "any";
+
+    act(() => {
+      document.body.appendChild(element);
+    });
+    expect(element.shadowRoot?.textContent).toMatchInlineSnapshot(
+      `"styles.shadow.css"`
+    );
+
+    act(() => {
+      document.body.removeChild(element);
+    });
   });
 });