+++ shared/form/src/FormStore.spec.ts
@@ -1,0 +1,246 @@
+import { FormStore } from "./FormStore";
+
+describe("FormStore", () => {
+  it("general", () => {
+    const store = new FormStore();
+
+    store.setField("a", {
+      name: "a",
+      label: "字段A",
+      validate: {
+        required: true,
+      },
+    });
+
+    const mockValidateFields = jest.fn();
+
+    const result1 = store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(1, true, [
+      { message: "字段A为必填项", type: "error" },
+    ]);
+
+    expect(result1).toBe(false);
+
+    store.setField("b", {
+      name: "b",
+      validate: {
+        required: true,
+      },
+    });
+
+    store.setInitValue({
+      a: "hello",
+    });
+
+    const result2 = store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(2, true, [
+      { message: "b为必填项", type: "error" },
+    ]);
+
+    expect(result2).toBe(false);
+
+    store.setInitValue({
+      b: "world",
+    });
+
+    const result3 = store.validateFields(mockValidateFields);
+
+    expect(store.getAllValues()).toEqual({ a: "hello", b: "world" });
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(3, false, {
+      a: "hello",
+      b: "world",
+    });
+    expect(result3).toEqual({ a: "hello", b: "world" });
+
+    store.setField("c", {
+      name: "c",
+      label: "校验测试字段",
+      validate: {
+        min: 5,
+        max: 10,
+        required: true,
+        pattern: "\\d+",
+      },
+    });
+
+    store.setInitValue({
+      a: null,
+      c: "12",
+    });
+
+    const result4 = store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(4, true, [
+      { message: "字段A为必填项", type: "error" },
+      { message: "校验测试字段至少包含 5 个字符", type: "error" },
+    ]);
+    expect(result4).toBe(false);
+
+    store.setInitValue({
+      a: 123,
+      c: "12345678901",
+    });
+
+    const result5 = store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(5, true, [
+      { message: "校验测试字段不能超过 10 个字符", type: "error" },
+    ]);
+    expect(result5).toBe(false);
+
+    store.setInitValue({
+      a: 123,
+      c: "abcdef",
+    });
+
+    store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(6, true, [
+      { message: "校验测试字段没有匹配正则 \\d+", type: "error" },
+    ]);
+
+    store.setInitValue({
+      c: "123456",
+    });
+
+    store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(7, false, {
+      a: 123,
+      b: "world",
+      c: "123456",
+    });
+
+    // not-render should skip
+    store.setField("d-notRender", {
+      name: "d-notRender",
+      notRender: true,
+      validate: {
+        required: true,
+      },
+    });
+
+    const result6 = store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(8, false, {
+      a: 123,
+      b: "world",
+      c: "123456",
+    });
+
+    expect(result6).toEqual({
+      a: 123,
+      b: "world",
+      c: "123456",
+    });
+
+    expect(store.getAllValues()).toEqual({
+      a: 123,
+      b: "world",
+      c: "123456",
+    });
+
+    // update notRender to false
+    store.setField("d-notRender", {
+      name: "d-notRender",
+      notRender: false,
+      validate: {
+        required: true,
+      },
+    });
+
+    store.setInitValue({
+      "d-notRender": "abc",
+    });
+
+    store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(9, false, {
+      a: 123,
+      b: "world",
+      c: "123456",
+      "d-notRender": "abc",
+    });
+
+    // not render to be true
+    store.setField("d-notRender", {
+      name: "d-notRender",
+      notRender: true,
+      validate: {
+        required: true,
+      },
+    });
+
+    // get Value should filter not-render item
+    expect(store.getAllValues()).toEqual({
+      a: 123,
+      b: "world",
+      c: "123456",
+    });
+
+    store.setField("validator-item", {
+      name: "validator-item",
+      label: "高级校验字段",
+      validate: {
+        validator: (value: any): string => (!value.name ? "名字不能为空" : ""),
+      },
+    });
+
+    store.setInitValue({
+      "validator-item": {
+        a: "1",
+      },
+    });
+
+    store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(10, true, [
+      {
+        message: "名字不能为空",
+        type: "error",
+      },
+    ]);
+
+    store.setInitValue({
+      "validator-item": {
+        name: "test-name",
+      },
+    });
+
+    store.validateFields(mockValidateFields);
+
+    expect(mockValidateFields).toHaveBeenNthCalledWith(11, false, {
+      a: 123,
+      b: "world",
+      c: "123456",
+      "validator-item": {
+        name: "test-name",
+      },
+    });
+
+    expect(store.getFieldsValue("a")).toBe(123);
+    expect(store.getFieldsValue()).toEqual({
+      a: 123,
+      b: "world",
+      c: "123456",
+      "validator-item": {
+        name: "test-name",
+      },
+    });
+
+    store.resetFields("validator-item");
+
+    expect(store.getAllValues()).toEqual({
+      a: 123,
+      b: "world",
+      c: "123456",
+    });
+
+    store.resetFields();
+
+    expect(store.getAllValues()).toEqual({});
+  });
+});